// tools/generators/domain/schema.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface DomainOptions {
  /**
   * Domain name
   */
  name: string;
  /**
   * Create buildable libraries
   */
  buildable?: boolean;
  /**
   * Skip babel
   */
  skipBabel?: boolean;
  [k: string]: unknown;
}